Allow arbitrary labels to be added
We want to be able to track things related to specific OKRs or ad-hoc plans. 
Milestones are great for this, but an issue can only be in one of those.
This way we can create labels on the fly, and we can remove them by hand once they're no longer valuable.

The githubLabelSync is great for making sure we have the minimum labels across our repos, but we it shouldn't be enforcing a maximum.